+++ TGL/src/collision.cpp
@@ -37,7 +37,7 @@
 
 
 
-bool collision_with_point(SDL_Surface *s1,float x1,float y1,int hot_x1,int hot_y1,int a1,float scale1,float x2,float y2)
+bool collision_with_point(SDL_Surface *s1,float x1,float y1,int hot_x1,int hot_y1,float a1,float scale1,float x2,float y2)
 {
 	// Radius check:
 	int collision_pixels=0;
@@ -124,7 +124,7 @@ bool collision_with_point(SDL_Surface *s1,float x1,float y1,int hot_x1,int hot_y
 
 
 
-bool collision(SDL_Surface *s1,float x1,float y1,int hot_x1,int hot_y1,int a1,float scale1,SDL_Surface *s2,float x2,float y2,int hot_x2,int hot_y2,int a2,float scale2)
+bool collision(SDL_Surface *s1,float x1,float y1,int hot_x1,int hot_y1,float a1,float scale1,SDL_Surface *s2,float x2,float y2,int hot_x2,int hot_y2,float a2,float scale2)
 {
 	// Radius check:
 	int collision_pixels=0;
@@ -246,7 +246,7 @@ bool collision(SDL_Surface *s1,float x1,float y1,int hot_x1,int hot_y1,int a1,fl
 
 
 
-bool collision_vector(SDL_Surface *s1,float x1,float y1,int hot_x1,int hot_y1,int a1,float scale1,SDL_Surface *s2,float x2,float y2,int hot_x2,int hot_y2,int a2,float scale2,float *collision_vector_x,float *collision_vector_y)
+bool collision_vector(SDL_Surface *s1,float x1,float y1,int hot_x1,int hot_y1,float a1,float scale1,SDL_Surface *s2,float x2,float y2,int hot_x2,int hot_y2,float a2,float scale2,float *collision_vector_x,float *collision_vector_y)
 {
 	List<int> collision_x,collision_y;	
 	int collision_pixels=0;
